backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.2",
    "fastapi>=0.135.1",
    "fastembed>=0.7.4",
    "httpx>=0.28.1",
    "langchain-groq>=0.3.0",
    "langgraph>=1.1.0",
    "mcp[cli]>=1.26.0",
    "opentelemetry-api>=1.40.0",
    "opentelemetry-exporter-otlp>=1.40.0",
    "opentelemetry-instrumentation-fastapi>=0.61b0",
    "opentelemetry-instrumentation-httpx>=0.61b0",
    "opentelemetry-sdk>=1.40.0",
    "prometheus-client>=0.24.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.17.0",
    "redis>=7.3.0",
    "sqlmodel>=0.0.24",
    "structlog>=25.5.0",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = ["black>=26.3.0", "mypy>=1.19.1", "pytest>=9.0.2", "ruff>=0.15.5"]
